Crate google-appsactivity1

This documentation was generated from appsactivity crate version 0.0.1. The original source code can be found on github.

Features

It seems there is nothing you can do here ... .

activities (list)

Everything else about the appsactivity v1 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.activities().list(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-appsactivity1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-appsactivity1" as appsactivity1;
use appsactivity1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use appsactivity1::Appsactivity;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Appsactivity::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.activities().list()
             .user_id("accusam")
             .source("takimata")
             .page_token("justo")
             .page_size(-1)
             .grouping_strategy("erat")
             .drive_file_id("labore")
             .drive_ancestor_id("sea")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Structs

Activity
An Activity resource is a combined view of multiple events. An activity has a list of individual events and a combined view of the common fields among all events.

This type is not used in any activity, and only used as part of another schema.

ActivityListMethodBuilder
Returns a list of activities visible to the current logged in user. Visible activities are determined by the visiblity settings of the object that was acted on, e.g. Drive files a user can see. An activity is a record of past events. Multiple events may be merged if they are similar. A request is scoped to activities from a given Google service using the source parameter.

ActivityMethodsBuilder
A builder providing access to all methods supported on activity resources. It is not used directly, but through the Appsactivity hub.

Appsactivity
Central instance to access all Appsactivity related resource activities

Event
Represents the changes associated with an action taken by a user.

This type is not used in any activity, and only used as part of another schema.

ListActivitiesResponse
The response from the list request. Contains a list of activities and a token to retrieve the next page of results.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list activities (response)

Move
Contains information about changes in an object's parents as a result of a move type event.

This type is not used in any activity, and only used as part of another schema.

Parent
Contains information about a parent object. For example, a folder in Drive is a parent for all files within it.

This type is not used in any activity, and only used as part of another schema.

Permission
Contains information about the permissions and type of access allowed with regards to a Google Drive object. This is a subset of the fields contained in a corresponding Drive Permissions object.

This type is not used in any activity, and only used as part of another schema.

PermissionChange
Contains information about a Drive object's permissions that changed as a result of a permissionChange type event.

This type is not used in any activity, and only used as part of another schema.

Photo
Photo information for a user.

This type is not used in any activity, and only used as part of another schema.

Rename
Contains information about a renametype event.

This type is not used in any activity, and only used as part of another schema.

Target
Information about the object modified by the event.

This type is not used in any activity, and only used as part of another schema.

User
A representation of a user.

This type is not used in any activity, and only used as part of another schema.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.