Module google_appsactivity1::api

Structs

Activity
An Activity resource is a combined view of multiple events. An activity has a list of individual events and a combined view of the common fields among all events.

ActivityListCall
Returns a list of activities visible to the current logged in user. Visible activities are determined by the visibility settings of the object that was acted on, e.g. Drive files a user can see. An activity is a record of past events. Multiple events may be merged if they are similar. A request is scoped to activities from a given Google service using the source parameter.

ActivityMethods
A builder providing access to all methods supported on activity resources. It is not used directly, but through the Appsactivity hub.

Appsactivity
Central instance to access all Appsactivity related resource activities

Event
Represents the changes associated with an action taken by a user.

ListActivitiesResponse
The response from the list request. Contains a list of activities and a token to retrieve the next page of results.

Move
Contains information about changes in an object's parents as a result of a move type event.

Parent
Contains information about a parent object. For example, a folder in Drive is a parent for all files within it.

Permission
Contains information about the permissions and type of access allowed with regards to a Google Drive object. This is a subset of the fields contained in a corresponding Drive Permissions object.

PermissionChange
Contains information about a Drive object's permissions that changed as a result of a permissionChange type event.

Photo
Photo information for a user.

Rename
Contains information about a renametype event.

Target
Information about the object modified by the event.

User
A representation of a user.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.