Struct google_appsactivity1::Event

pub struct Event {
    pub rename: Option<Rename>,
    pub target: Option<Target>,
    pub additional_event_types: Option<Vec<String>>,
    pub move_: Option<Move>,
    pub permission_changes: Option<Vec<PermissionChange>>,
    pub user: Option<User>,
    pub event_time_millis: Option<String>,
    pub primary_event_type: Option<String>,
    pub from_user_deletion: Option<bool>,
}

Represents the changes associated with an action taken by a user.

This type is not used in any activity, and only used as part of another schema.

Fields

rename

Extra information for rename type events, such as the old and new names.

target

Information specific to the Target object modified by the event.

additional_event_types

Additional event types. Some events may have multiple types when multiple actions are part of a single event. For example, creating a document, renaming it, and sharing it may be part of a single file-creation event.

move_

Extra information for move type events, such as changes in an object's parents.

permission_changes

Extra information for permissionChange type events, such as the user or group the new permission applies to.

user

Represents the user responsible for the event.

event_time_millis

The time at which the event occurred formatted as Unix time in milliseconds.

primary_event_type

The main type of event that occurred.

from_user_deletion

Whether this event is caused by a user being deleted.